requirements.txt
torch==1.12.1
torchvision==0.13.1
Pillow==9.0.1
numpy==1.21.2
matplotlib==3.4.3
scikit-learn==1.0.1
tqdm==4.62.3
opencv-python==4.5.3.20210927
pandas==1.3.3